fix: keep appearance tests within file budget
